backend/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
ipython
traitlets
gunicorn
